requirements.txt
fastapi
uvicorn
sqlalchemy
oracledb
apscheduler
python-dotenv
httpx
holidays
openpyxl
reportlab
